backend/pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Deep Research Agent with Multi-Agent Workflow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework>=1.0.0b251112.post1",
    "arxiv>=2.3.1",
    "fastapi>=0.121.2",
    "google-api-python-client>=2.187.0",
    "openai>=2.8.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
